users/requirements.txt
Flask==1.1.2
requests==2.25.1
PyJWT==2.1.0
pymongo==3.11.4
bcrypt==3.2.0